requirements.txt
APScheduler==3.6.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
cn2an==0.4.1
h11==0.14.0
idna==3.7
itchat-uos==1.5.0.dev0
numpy==1.26.4
outcome==1.3.0.post0
pandas==2.2.2
pillow==10.3.0
pypng==0.20220715.0
PyQRCode==1.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzbar==0.1.8
requests==2.31.0
retrying==1.3.3
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
suning-sdk==1.0.0
tqdm==4.66.4
trio==0.25.0
trio-websocket==0.11.1
typing-extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wsproto==1.2.0
